Made `checkout_branches` script overridable.
Was necessary to do this so that the same script could be reused for the release branch builds. The release branch builds differ from the feature branch the commit hashes to build are known whereas a feature branch build is triggered from a PR on any of the 3 development repos and involves deducing which PR triggered it and what the commit hashes are.