Changed repo name to BJSS_liveobs_client_modules.
Whenever the repo is cloned it will be put in a directory called `BJSS_liveobs_client_modules` unless otherwise specified. Making the directory name differ from the repo name just means having to remember to do that everywhere rather than letting it fall to the default. At the moment there are only 2 GoCD jobs and our local environments to have to remember to do this, but that number will increase and cause time lost when we don't remember to do this with other things we are working on.